+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Console Logger
 - Composite Logger
+- Logger Factory
 
 ## [0.4.1] - 2024-10-26